lib: disallow direct access to global or global properties in builtins
This patch makes the global proxy object available to per_context.js,
bootstrappers, and native modules through a parameter to
ScriptCompiler::CompileFunctionInContext(), and name it `globalProxy`.

It disallows access to `global` or any property on the global
object in the native modules by removing `node: true` from `env` in
the eslint configuration for the `lib` directory, hence limiting the
Node.js-specific global variables that can be accessed by the native
modules to a known set in `lib/.eslintrc.yaml`.
Access to global properties such as `setTimeout` in native modules
should be done by explicitly referencing them as
`globalProxy.setTimeout` instead of implicitly falling back to a
lookup from the global object.
It's still possible to directly reference the globals and `global`
in the native modules when they are compiled and run locally,
but such references can now be caught by the linter.

This allows us to postpone `globalProxy.global = globalProxy` to
a later point in bootstrap and moves that from C++ to JS.