src: snapshot Environment upon instantiation
Note that all the AliasedBuffer on the heap are not deserialized
from the snapshot but are instead created from scratch (with 0
as values). This currently works under the assumption that
at the point where the snapshot is taken, no AliasedBuffer are
modified, so even if we effectively discard the content of these
buffers in the snapshot and reinitialize them all to 0, the
state is still consistent.

This, however, is not going to work once we start modifying
these buffers during bootstrap, but currently we do not.