src: switch BindingData back to v8::External
This unblocks snapshotting, because we cannot refer from the
`ObjectTemplate`s and `FunctionTemplate`s to context-specific
data such as binding objects.

This is a bit unfortunate. Reading Externals currently does not
match the performance of retrieving fields from binding objects,
and it would be nice to have more than a single internal field
available. It also forces us to know the memory address of the
`External` target before the `Isolate` is created when
creating or reading from a snapshot.

Ideally, this will be addressed by upstream V8 in some way.