vm: fix vm.compileFunction() dynamic import callback leak
This fixes the leak in vm.compileFunction() when
importModuleDynamically is set by using a symbol in the host
defined options to maintain the lifetime of the module referer object.
This relies on the currenly stage-3 proposal that makes it possible
to store symbols in the weak map as keys because otherwise we
can't find a primitive that we can use to keep the referer object
alive in a GC-aware fashion.