src: add internal id to generate default snapshot as C++ code
This adds an internal "node:default_snapshot_main" that can be
used to generate the default snapshot as C++ code, so that
we can use the output to check the reproducibility of snapshots.

To make the snapshots more human-readable the snapshot generated
from this path would encode the snapshot data into number literals.
The snapshot generated for actual compilation would still use
string literals if NODE_MKSNAPSHOT_USE_STRING_LITERALS is
set.