src: unmake DirHandle an AsyncWrap
DirHandle does not actually follow the lifecycle of an AsyncWrap,
it only emits the init event and its ID is not connected to
other async operations. The actual async provider is
FSReqCallback and the DirHandle is just a wrapper without
anything asynchronous, the same way that net.Socket etc. are not
async providers but mere wrappers. Simplify it by unmaking it
an AsyncWrap.