src: only block on user blocking worker tasks
Previously, we block on all worker tasks in the main thread,
which can cause a deadlock if the worker task post another
foreground task and wait for it to complete. This patch changes
it to only block on worker tasks that have kUserBlocking priority
to avoid the deadlock.

In addition this updates the task queue to be a priority queue
and order the task based on priority, so that higher priority
tasks get executed first, as the documentation suggests.